+++ app/src/main/java/ly/count/android/demo/MainActivity.java
@@ -30,12 +30,16 @@ public class MainActivity extends Activity {
         activity = this;
         Context appC = getApplicationContext();
 
+        HashMap<String, String> customHeaderValues = new HashMap<>();
+        customHeaderValues.put("foo", "bar");
+
         Countly.onCreate(this);
         Countly.sharedInstance().setLoggingEnabled(true);
         Countly.sharedInstance().enableCrashReporting();
         Countly.sharedInstance().setViewTracking(true);
         Countly.sharedInstance().setAutoTrackingUseShortName(true);
         Countly.sharedInstance().setRequiresConsent(true);
+        Countly.sharedInstance().addCustomNetworkRequestHeaders(customHeaderValues);
         Countly.sharedInstance().setRemoteConfigAutomaticDownload(true, new RemoteConfig.RemoteConfigCallback() {
             @Override
             public void callback(String error) {